Update Makefile to use RELEASE_VER for rpm version
This change is to allow this git repo, to honor a
RELEASE_VER file for RPM naming convetion. This will be
used over counting the number of commits since the last
v<major>.<minor> tag in the repo.

Mostly this is added to allow internal and external
rpm naming syncronization with Intel IFS releases for OPA.